requirements.txt
pandas
tweepy
moviepy
python-dotenv
beautifulsoup4
lxml
curl_cffi
mysql-connector-python